+++ include/generalized_information_filter/common.h
