+++ include/generalized_information_filter/filter.h
@@ -1,8 +1,8 @@
 #ifndef GIF_FILTER_HPP_
 #define GIF_FILTER_HPP_
 
-#include "generalized_information_filter/BinaryResidual.hpp"
-#include "generalized_information_filter/common.hpp"
+#include "generalized_information_filter/binary-residual.h"
+#include "generalized_information_filter/common.h"
 
 namespace GIF {
